requirements.txt
asgiref==3.9.1
Django==5.2.5
django-taggit==6.1.0
django-widget-tweaks==1.5.0
pillow==11.3.0
sqlparse==0.5.3
